+++ app/config/scripts/nginx.sh
@@ -15,6 +15,10 @@ chown -Rf www-data:www-data /code
 chown -Rf www-data:www-data /media
 chown -Rf www-data:www-data /php
 
+# Emulate /srv/binding
+mkdir -p /srv/bindings
+ln -s / "$HOME" || true
+
 # Set up our certs for the appserver with nginx
 if [ ! -f "/certs/appserver.pem" ]; then
   openssl genrsa -out /certs/appserver.key 2048 && \